requirements.txt
instabot==0.60.0
requests==2.22.0
python-dotenv==0.10.3

